rearrangement pytest
