+++ brainevent/_op/kernix_cuda_test.py
@@ -19,7 +19,6 @@ import pytest
 import numpy as np
 
 from brainevent._test_util import requires_gpu
-from brainevent._op import kernix_toolchain as kt
 
 VADD_CUDA_SRC = r"""
 #include <cuda_runtime.h>
@@ -178,143 +177,3 @@ def test_two_outputs(multi_out_module):
     np.testing.assert_allclose(
         np.asarray(hi), np.arange(n // 2, n, dtype=np.float32)
     )
-
-
-# ---------------------------------------------------------------------------
-# Compute-capability resolution tests
-# ---------------------------------------------------------------------------
-
-@pytest.mark.parametrize("raw,expected", [
-    ("8.6", "sm_86"), ("86", "sm_86"), ("sm_86", "sm_86"),
-    ("compute_86", "sm_86"), (" 8.6 ", "sm_86"), ("9.0a", "sm_90a"),
-    ("90a", "sm_90a"), ("12.0", "sm_120"), ("120", "sm_120"),
-])
-def test_normalize_arch_ok(raw, expected):
-    assert kt.normalize_arch(raw) == expected
-
-
-@pytest.mark.parametrize("bad", ["", "   ", "abc", "x", "8", "sm_", ".."])
-def test_normalize_arch_bad(bad):
-    with pytest.raises(ValueError):
-        kt.normalize_arch(bad)
-
-
-def test_resolve_explicit_wins(monkeypatch):
-    monkeypatch.setattr(kt, "_arch_from_jax", lambda: ["sm_99"])
-    assert kt.resolve_compute_capabilities("8.6") == ["sm_86"]
-    assert kt.resolve_compute_capabilities(["8.6", "9.0"]) == ["sm_86", "sm_90"]
-
-
-def test_resolve_explicit_comma_string(monkeypatch):
-    monkeypatch.setattr(kt, "_arch_from_jax", lambda: ["sm_99"])
-    assert kt.resolve_compute_capabilities("8.0,8.6") == ["sm_80", "sm_86"]
-    assert kt.resolve_compute_capabilities(" 8.0 , , 8.6 ") == ["sm_80", "sm_86"]
-    assert kt.resolve_compute_capabilities(["8.0,8.6", "9.0"]) == [
-        "sm_80", "sm_86", "sm_90"]
-
-
-def test_resolve_precedence_config_over_env(monkeypatch):
-    monkeypatch.setenv("BRAINEVENT_COMPUTE_CAPABILITIES", "8.0")
-    monkeypatch.setattr(kt, "_arch_from_jax", lambda: ["sm_99"])
-    kt.set_compute_capabilities("8.6")
-    try:
-        assert kt.resolve_compute_capabilities() == ["sm_86"]
-    finally:
-        kt.set_compute_capabilities(None)
-
-
-def test_resolve_env_over_jax(monkeypatch):
-    monkeypatch.setenv("BRAINEVENT_COMPUTE_CAPABILITIES", " 8.0 , , 8.6 ")
-    monkeypatch.setattr(kt, "_arch_from_jax", lambda: ["sm_99"])
-    assert kt.resolve_compute_capabilities() == ["sm_80", "sm_86"]
-
-
-def test_resolve_jax_over_smi(monkeypatch):
-    monkeypatch.delenv("BRAINEVENT_COMPUTE_CAPABILITIES", raising=False)
-    monkeypatch.setattr(kt, "_arch_from_jax", lambda: ["sm_86"])
-    monkeypatch.setattr(kt, "_arch_from_nvidia_smi", lambda: ["sm_70"])
-    assert kt.resolve_compute_capabilities() == ["sm_86"]
-
-
-def test_resolve_raises_when_all_absent(monkeypatch):
-    from brainevent._error import GpuArchDetectionError
-    monkeypatch.delenv("BRAINEVENT_COMPUTE_CAPABILITIES", raising=False)
-    monkeypatch.setattr(kt, "_arch_from_jax", lambda: None)
-    monkeypatch.setattr(kt, "_arch_from_nvidia_smi", lambda: None)
-    with pytest.raises(GpuArchDetectionError):
-        kt.resolve_compute_capabilities()
-
-
-def test_gencode_single():
-    assert kt.gencode_flags(["sm_86"]) == [
-        "-gencode", "arch=compute_86,code=sm_86",
-        "-gencode", "arch=compute_86,code=compute_86",
-    ]
-
-
-def test_gencode_multi_ptx_for_highest():
-    out = kt.gencode_flags(["sm_80", "sm_90", "8.6"])
-    assert "arch=compute_80,code=sm_80" in out
-    assert "arch=compute_86,code=sm_86" in out
-    assert "arch=compute_90,code=sm_90" in out
-    assert out[-1] == "arch=compute_90,code=compute_90"
-
-
-def test_gencode_empty_raises():
-    with pytest.raises(ValueError):
-        kt.gencode_flags([])
-
-
-def test_cache_uses_platform_ext(tmp_path, monkeypatch):
-    from brainevent._op import kernix_cache as kcache
-    from brainevent._op import kernix_toolchain as ktool
-    monkeypatch.setattr(ktool.sys, "platform", "win32")
-    c = kcache.CompilationCache(base_dir=str(tmp_path))
-    assert c.lookup("m", "deadbeef") is None
-    src = tmp_path / "src.dll"
-    src.write_bytes(b"x")
-    dest = c.store("m", "deadbeef", str(src))
-    assert dest.name == "m.dll"
-    assert c.lookup("m", "deadbeef") == dest
-
-
-def test_cache_store_atomic_move(tmp_path):
-    from brainevent._op import kernix_cache as kcache
-    c = kcache.CompilationCache(base_dir=str(tmp_path))
-    build = tmp_path / "build"
-    build.mkdir()
-    so = build / "m.so"
-    so.write_bytes(b"hello")
-    dest = c.store("m", "k", str(so))
-    assert dest.read_bytes() == b"hello"
-    assert not so.exists()
-
-
-def test_config_set_compute_capability():
-    import brainevent
-    from brainevent._op import kernix_toolchain as ktool
-    brainevent.config.set_compute_capability("8.6")
-    try:
-        assert brainevent.config.get_compute_capability() == ["sm_86"]
-        assert ktool.resolve_compute_capabilities() == ["sm_86"]
-    finally:
-        brainevent.config.set_compute_capability(None)
-    assert brainevent.config.get_compute_capability() is None
-
-
-def test_set_compute_capabilities_comma_string():
-    kt.set_compute_capabilities("8.6,8.0")
-    try:
-        assert kt.get_compute_capabilities() == ["sm_86", "sm_80"]
-    finally:
-        kt.set_compute_capabilities(None)
-    assert kt.get_compute_capabilities() is None
-
-
-def test_config_set_compute_capability_comma():
-    import brainevent
-    brainevent.config.set_compute_capability("8.6, 8.0")
-    try:
-        assert brainevent.config.get_compute_capability() == ["sm_86", "sm_80"]
-    finally:
-        brainevent.config.set_compute_capability(None)
