+++ brainevent/_op/numba_cuda_ffi_test.py
@@ -26,6 +26,9 @@ import jax.numpy as jnp
 import numpy as np
 import pytest
 
+from brainevent._op import numba_cuda_ffi
+from brainevent._op.numba_cuda_ffi import _compute_launch_config
+
 # Check if Numba CUDA is available
 numba_installed = importlib.util.find_spec('numba') is not None
 numba_cuda_available = False
@@ -39,6 +42,10 @@ if numba_installed:
 
 gpu_platform = jax.default_backend() == 'gpu'
 numba_cuda_available = numba_cuda_available and gpu_platform
+
+requires_gpu = pytest.mark.skipif(
+    not numba_cuda_available, reason="Numba CUDA not available"
+)
 
 
 @pytest.mark.skipif(not numba_cuda_available, reason="Numba CUDA not available")
@@ -573,14 +580,14 @@ class TestNumbaCudaKernelXLAStream(unittest.TestCase):
 
 
 # ===========================================================================
-# 1. Validation / Error Handling (no GPU needed)
+# 1. Validation / Error Handling
 # ===========================================================================
 
 
 class TestNumbaCudaCallableValidation(unittest.TestCase):
-    """Validation tests that do not require a GPU."""
+    """Validation tests for the callable wrapper."""
 
-    @pytest.mark.skipif(not numba_cuda_available, reason="Numba not installed")
+    @pytest.mark.skipif(not numba_cuda_available, reason="Numba CUDA not available")
     def test_non_callable_raises_typeerror(self):
         from brainevent import numba_cuda_callable
 
@@ -590,7 +597,7 @@ class TestNumbaCudaCallableValidation(unittest.TestCase):
                 outs=jax.ShapeDtypeStruct((10,), jnp.float32),
             )
 
-    @pytest.mark.skipif(not numba_cuda_available, reason="Numba not installed")
+    @pytest.mark.skipif(not numba_cuda_available, reason="Numba CUDA not available")
     def test_returns_callable(self):
         from brainevent import numba_cuda_callable
 
@@ -1547,68 +1554,8 @@ class TestNumbaCudaCallableDtypes(unittest.TestCase):
 
 
 # ---------------------------------------------------------------------------
-# Merged audit regression tests
+# Audit regression tests
 # ---------------------------------------------------------------------------
-
-# Copyright 2025 BrainX Ecosystem Limited. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-
-"""Reproduction tests for the numba-CUDA FFI bridge audit fixes.
-
-Covers the GPU-bridge findings from ``dev/2026-06-13-op-issues.md``:
-
-* C1 - a raising kernel/callable surfaces as a JAX error, not silent success.
-* C2 - ``float16`` round-trips; ``bfloat16`` (unsupported by numba CUDA) is
-  rejected loudly instead of mis-decoded.
-* M3 - ``_compute_launch_config`` guards zero/negative extents (no
-  ``ZeroDivisionError``); an empty launch is skipped without crashing.
-* H1 - repeated calls with an identical signature reuse one FFI registration.
-
-The pure-function ``_compute_launch_config`` tests run on any platform; the
-end-to-end tests require a real GPU and are skipped otherwise.
-"""
-
-import importlib.util
-import os
-import unittest
-
-os.environ['JAX_TRACEBACK_FILTERING'] = 'off'
-
-import jax
-import jax.numpy as jnp
-import numpy as np
-import pytest
-
-from brainevent._op import numba_cuda_ffi
-from brainevent._op.numba_cuda_ffi import _compute_launch_config
-
-# --- backend / numba-cuda detection ----------------------------------------
-numba_installed = importlib.util.find_spec('numba') is not None
-numba_cuda_available = False
-if numba_installed:
-    try:
-        from numba import cuda
-
-        numba_cuda_available = cuda.is_available()
-    except ImportError:
-        pass
-numba_cuda_available = numba_cuda_available and (jax.default_backend() == 'gpu')
-
-requires_gpu = pytest.mark.skipif(
-    not numba_cuda_available, reason='Numba CUDA / GPU backend not available'
-)
 
 
 class TestComputeLaunchConfig(unittest.TestCase):
